+++ plotoptix/__init__.py
@@ -1,5 +1,5 @@
 """
-3D ray tracing package for Python. Based on NVIDIA OptiX framework,
+3D ray tracing package for Python. Based on NVIDIA OptiX 7 framework,
 wrapped in RnD.SharpOptiX C#/C++/CUDA libraries by R&D Team.
 
 Copyright (C) 2019 R&D Team. All Rights Reserved.
@@ -12,8 +12,8 @@ __all__ = ["enums", "materials", "utils", "npoptix", "tkoptix"]
 
 __author__  = "Robert Sulej, R&D Team <dev@rnd.team>"
 __status__  = "beta"
-__version__ = "0.5.2"
-__date__    = "15 Oct 2019"
+__version__ = "0.7.0"
+__date__    = "17 Mar 2020"
 
 import logging