+++ plotoptix/materials.py
@@ -8,16 +8,16 @@ from plotoptix.enums import RtFormat
 __pkg_dir__ = os.path.dirname(__file__)
 
 m_flat = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::flat_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ]
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__flat",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
     }
 """
 Super-fast material, color is not shaded anyhow. Use color components range <0; 1>.
 """
 
 m_eye_normal_cos = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::cos_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ]
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__cos",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
     }
 """
 Fast material, color is shaded by the cos(eye-hit-normal). Use color components range
@@ -25,9 +25,9 @@ Fast material, color is shaded by the cos(eye-hit-normal). Use color components 
 """
 
 m_diffuse = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::diffuse_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 2 }
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__diffuse",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 2 }
     }
 """
 Lambertian diffuse material. Note it is available by default under the name "diffuse".
@@ -35,10 +35,10 @@ Use color components range <0; 1>.
 """
 
 m_matt_diffuse = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::diffuse_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 2 },
-      "VarFloat": { "surface_roughness": 1 }
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__diffuse",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 2 },
+      "VarFloat": { "base_roughness": 1 }
     }
 """
 Oren-Nayar diffuse material. Surface roughness range is <0; inf), 0 is equivalent to
@@ -46,9 +46,9 @@ the Lambertian "diffuse" material.
 """
 
 m_mirror = {
-      "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
-      "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
-      "VarInt": { "material_flags": 6 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__reflective",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 6 },
       "VarFloat3": {
         "surface_albedo": [ 1.0, 1.0, 1.0 ]
       }
@@ -61,10 +61,10 @@ range <0; 1>), which results with colorized reflections.
 """
 
 m_metalic = {
-      "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
-      "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
-      "VarInt": { "material_flags": 6 },
-      "VarFloat": { "surface_roughness": 0.002 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__reflective",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 6 },
+      "VarFloat": { "base_roughness": 0.002 },
 }
 """
 Strongly reflective, metalic material. Note, this material has default values: ``reflectivity_index = 1``
@@ -74,9 +74,9 @@ Roughness of the surface should be usually small.
 """
 
 m_plastic = {
-      "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
-      "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
-      "VarInt": { "material_flags": 6 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__reflective",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 6 },
       "VarFloat": {
         "reflectivity_index": 0.0,
         "reflectivity_range": 0.5,
@@ -93,13 +93,13 @@ reflectivity_range value (down to 0). Higher refraction_index gives a more gloss
 """
 
 m_matt_plastic = {
-      "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
-      "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
-      "VarInt": { "material_flags": 6 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__reflective",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 6 },
       "VarFloat": {
         "reflectivity_index": 0.0,
         "reflectivity_range": 0.5,
-        "surface_roughness": 0.001
+        "base_roughness": 0.001
       },
       "VarFloat3": {
         "refraction_index": [ 2.0, 2.0, 2.0 ],
@@ -111,9 +111,9 @@ Similar to :attr:`plotoptix.materials.m_plastic` but slightly rough surface.
 """
 
 m_clear_glass = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 12 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__glass",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 12 },
       "VarFloat": {
         "radiation_length": 0.0,
         "vol_scattering": 1.0,
@@ -130,14 +130,14 @@ and the color range is <0; inf>.
 """
 
 m_matt_glass = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 12 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__glass",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 12 },
       "VarFloat": {
         "radiation_length": 0.0,
         "vol_scattering": 1.0,
         "light_emission": 0.0,
-        "surface_roughness": 0.2
+        "base_roughness": 0.2
       },
       "VarFloat3": {
         "refraction_index": [ 1.4, 1.4, 1.4 ],
@@ -150,9 +150,9 @@ Glass with surface roughness configured to obtain matt appearance. Color compone
 """
 
 m_dispersive_glass = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 12 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__glass",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 12 },
       "VarFloat": {
         "radiation_length": 0.0,
         "vol_scattering": 1.0,
@@ -170,9 +170,9 @@ and the range is <0; inf>.
 """
 
 m_thin_walled = {
-      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
-      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
-      "VarInt": { "material_flags": 44 },
+      "RadianceProgram": "materials7.ptx::__closesthit__radiance__glass",
+      "OcclusionProgram": "materials7.ptx::__closesthit__occlusion",
+      "VarUInt": { "flags": 44 },
       "VarFloat": {
         "radiation_length": 0.0,
         "vol_scattering": 1.0,